Create base class factory for component state.
